% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacircle.R
\name{areacircle}
\alias{areacircle}
\title{Area of a circle}
\usage{
areacircle(x, plot_it = FALSE)
}
\arguments{
\item{x}{The radius of a circle.}

\item{plot_it}{Display a plot of \code{x} vs the output. Use logical.
\code{FALSE} by default.}
}
\value{
The area of a circle with radius \code{x}.
}
\description{
This function calculates the area of a circle.
}
\examples{
areacircle(5)
areacircle(2, plot_it=TRUE)
}
