% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument.R
\name{argument}
\alias{argument}
\title{Argument of a complex number}
\usage{
argument(r, i = 0)
}
\arguments{
\item{r}{Real part}

\item{i}{Complex part}
}
\value{
A complex number looks like \code{r}+\code{i}i.
}
\description{
Gives the angle of a complex number
}
\details{
We can get the argument using the arcTangent!
}
\examples{
argument(3,4)
argument(1)
}
