% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_cox_two_param.R
\name{Box_cox_two_param}
\alias{Box_cox_two_param}
\title{Two parameter Box-Cox Transform Function}
\usage{
Box_cox_two_param(x, lambda1 = 0, lambda2 = 0, change_neg_to_pos = 0)
}
\arguments{
\item{x}{Vector to be transformed to some power using the two parameter box-cox function.}

\item{lambda1}{Power to divide and raise \code{x+lambda2} by. It defaults to zero but could be a vector. This has to be a single value for the function to work.}

\item{lambda2}{A shift paramer which also defaults to zero. It should be a single value as well.}

\item{change_neg_to_pos}{makes all \code{x} positive by the addition of a constant value. The deefault is 0.}
}
\value{
A vector with a transformed value. It returns \code{log(x + lambda2)} for lambda1 equals 0, and \code{((x+lambda2)^lambda1 - 1) / lambda1} for lambda1 non-zero.
}
\description{
#'#' This function gives the Box-Cox transformation for two parameters from a given variable. I'm making it internal,
by not \code{export}ing, because I don't want users to use it.
Why not? Because I said so!
}
