% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.inv.R
\name{boxcox.inv}
\alias{boxcox.inv}
\title{Boxcox inverses of a vector}
\usage{
boxcox.inv(x, lambda)
}
\arguments{
\item{x}{The vector to be boxcox inversed.}

\item{lambda}{The parameter of Box–Cox transformation}
}
\value{
A vector that is the boxcox inversed transformation of \code{x}.
}
\description{
Inverse transforms via the Box-Cox transform.
}
\details{
We only do the boxcox transformations that only requires one input.

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{boxcox.inv} function only requires one input.
     \item There's nothing else to say about \code{boxcox.inv}.
}
}
\examples{
boxcox.inv(1:10, 2)
boxcox.inv(2, 0)
}
