% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxCox.R
\name{boxCox}
\alias{boxCox}
\title{One Parameter Box-Cox Tranform Function}
\usage{
boxCox(x, lambda = 0)
}
\arguments{
\item{x}{the vector used in Box-Cox transformation}

\item{lambda}{the parameter that is estimated using the profile likelihood function}
}
\value{
The vector after tranforming \code{x} using \code{lambda} doing box-cox transformation
}
\description{
This function calculates the one-parameter Box-Cox transformation:
}
