% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxCox.R
\name{boxCox2}
\alias{boxCox2}
\title{Two Parameters Box-Cox Tranform Function}
\usage{
boxCox2(x, y = 0, lambda = 0)
}
\arguments{
\item{x}{the first parameter used in Box-Cox transformation}

\item{y}{the parameter that is estimated using the profile likelihood function}

\item{z}{the second parameter used in Box-Cox transformation}
}
\value{
The vector after tranforming \code{x} and \code{y} using \code{lambda} doing box-cox transformation.
}
\description{
This function calculates the two-parameters Box-Cox transformations:
}
