% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{clean_characters}
\alias{clean_characters}
\title{Title of the Function}
\usage{
clean_characters(df, convert_real_na = TRUE, trim = TRUE, extra_bad = NULL)
}
\arguments{
\item{x}{Description of the first argument XX.}
}
\value{
Description of the return value ZZ.
}
\description{
A short description of what the function does OK.
}
\examples{
df1 <- data.frame(
 a = c("hello", "  NA", "   "),
 b = c("x", NA , "y"),
 c = c(1,2,NA),
 stringsAsFactors = FALSE
 )

clean_characters(df1)

df2 <- df1 |>
 dplyr::mutate(aa = clean_characters(a))
df2

a = c("hello", "hi", NA)
b = c(1,2)

clean_characters(a)
clean_characters(b)

clean_characters(df1[1:3])
clean_characters(df1$a)

}
