% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic.R
\name{cubic}
\alias{cubic}
\title{Squares numeric data}
\usage{
cubic(x)
}
\arguments{
\item{x}{numeric, vector}
}
\value{
numeric, vector
}
\description{
Squares numeric data (single a number, or a numeric vector)
}
\examples{
cubic(5)
cubic(0)
cubic(c(1,2,3))
}
