% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff}
\alias{diff}
\title{Calculate the difference between two numbers}
\usage{
diff(x, y)
}
\arguments{
\item{x}{The number that \code{y} is subtracted from.}

\item{y}{The number that is substracted from \code{x}.}
}
\value{
The difference between \code{x} and \code{y}.
}
\description{
Calculate the difference between two numbers
}
\details{
This function caluculated the diffence between two numbers that are inputted by the user (\code{x},\code{y}).
}
\examples{
diff(1,5)
diff(100,3)
}
