% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expform.R
\name{expform}
\alias{expform}
\title{Exponential form}
\usage{
expform(r, i = 0)
}
\arguments{
\item{r}{Real part}

\item{i}{Complex part}
}
\value{
A complex number looks like \code{r}+\code{i}i.
}
\description{
Gives the exponential form of a complex number
}
\details{
We can get the exponential form using the boht the modulus and argument!
}
\examples{
expform(3,4)
expform(1)
}
