% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R, R/reciprocal.R, R/square.R
\name{cube}
\alias{cube}
\alias{reciprocal}
\alias{square}
\title{Raise a vector to a specific power}
\usage{
cube(x)

reciprocal(x)

square(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A vector containing the computed input
}
\description{
Raise a vector to a specific power
}
\examples{
square(5)
square(1:5)
cube(3)
cube(1:9)
}
