% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.na.R
\name{filter.na}
\alias{filter.na}
\title{Filter out the observations/features that contains the NA values within data.frame}
\usage{
filter.na(x, direction)
}
\arguments{
\item{x}{The dataframe needs to be checked.}

\item{direction}{user specify how the NA observations should be deleted,
by "row" or "col".}
}
\value{
A list with the first elemenet indicating whether ther is any missing values in the dataframe,
second element indicating the dataset with NA values omitted.
}
\description{
Tells whether the data input contains any NA values,
and filter out the NA values by user's specification - namely, either row or columne.
}
\details{
We only do the boxcox transformations that only requires one input.

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{boxcox} function only requires one input.
     \item There's nothing else to say about \code{boxcox}.
}
}
\examples{
m <- as.data.frame(cbind(c(1,2,1), c(2,2,NA)))
filter.na(m, direction = "col")
filter.na(m, direction = "row")$FilteredData

}
