% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housingdata.R
\docType{data}
\name{housingdata}
\alias{housingdata}
\title{Artificiant Housing Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 5 columns.
}
\usage{
housingdata
}
\description{
Includes made up data!
}
\keyword{datasets}
