% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxCoxInverse.R
\name{invBoxCox}
\alias{invBoxCox}
\title{Box-Cox Tranform Inverse Function}
\usage{
invBoxCox(x, lambda = 0)
}
\arguments{
\item{x}{the first parameter used in Box-Cox trannsformation}

\item{lambda}{the parameter that is estimated using the profile likelihood function}
}
\value{
The vector after inverse the transformation of \code{x} using \code{lambda}.
}
\description{
This function calculates the Box-Cox transformation inverse:
}
