% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_if_possible.R
\name{is_list_numeric_singletons}
\alias{is_list_numeric_singletons}
\title{Check if a list contains numeric singletons}
\usage{
is_list_numeric_singletons(x)
}
\arguments{
\item{x}{List to investigate}
}
\value{
Single logical; \code{TRUE} if all list entries are numeric
vectors of length 1.
}
\description{
Internal function.
}
