% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{logarithm}
\alias{logarithm}
\title{Take the logarithm of a vector}
\usage{
logarithm(x, base = 10)
}
\arguments{
\item{x}{The vector to be logged}

\item{base}{The base of the log}
}
\value{
A vector that is the log of \code{x}.

The base of the log to be applied to \code{x}
}
\description{
That's it -- this function just logs a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{lologarithm} function is quite simple.
     \item There's nothing else to say about \code{logarithm}.
}
}
\examples{
logarithm(1:10)
logarithm(-5)
}
