% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus.R
\name{modulus}
\alias{modulus}
\title{Modulus of a complex number}
\usage{
modulus(r, i = 0)
}
\arguments{
\item{r}{Real part}

\item{i}{Complex part}
}
\value{
A complex number looks like \code{r}+\code{i}i.
}
\description{
Gives the magnitude of a complex number
}
\details{
We can get the modulus using the Euclidean norm!
}
\examples{
modulus(3,4)
modulus(1)
}
