% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Motor Trend Car Road Tests}
\format{
An object of class \code{data.frame} with 32 rows and 11 columns.
}
\source{
Henderson and Velleman (1981), Building multiple regression models interactively. Biometrics, 37, 391–411.
}
\usage{
mtcars
}
\description{
The data was extracted from the 1974 Motor Trend US magazine,
and comprises fuel consumption and 10 aspects of automobile design and
performance for 32 automobiles (1973–74 models).
}
\keyword{datasets}
