% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negtenvec.R
\docType{data}
\name{negtenvec}
\alias{negtenvec}
\title{Integer vector from -10 to -1}
\format{Numeric vector.}
\source{
This data was built from the R function call `seq(-10,-1,1)`.
}
\usage{
negtenvec
}
\description{
Self-explanatory. It is simply a vector of numbers from -10 to -1, useful for testing the powers library.
}
\seealso{
\code{\link{pwr}}, \code{\link{square}}, \code{\link{cube}}, \code{\link{reciprocal}}, \code{\link{tenvec}}
}
\keyword{datasets}
