% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig_operations.R
\name{polar_transform}
\alias{polar_transform}
\title{Transform Cartesian coordinates into polar coordinates}
\usage{
polar_transform(x, y)
}
\arguments{
\item{x}{A number or vector}

\item{y}{A number or vector}
}
\value{
a list with (1) the first element containing the radii and (2) the second element containing the angles.
}
\description{
Transform Cartesian coordinates into polar coordinates
}
\details{
This function transforms Cartesian coordinates into polar coordinates, as specified by \code{r} and \code{theta}.

\code{r} is the radius.
\code{theta} is the angle from the x=axis.

Definition of transformation from Cartesian to polar coodinates:
\itemize{
     \item \code{r = sqrt(x^2 + y^2)}.
     \item \code{theta = arctan(y/x)}.
}
}
\examples{
pola_transform(x=c(1,4,10), y=c(2,8,32))
pola_transform(x=c(-1,-4,-10), y=c(2,8,32))
}
