% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow2}
\alias{pow2}
\title{This is intended as an internal function
not recommended for the average user}
\usage{
pow2(x, a = 2)
}
\description{
This is intended as an internal function
not recommended for the average user
}
