% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_datatable_multi.R
\name{power_datatable}
\alias{power_datatable}
\title{Raise to power with datatable}
\usage{
power_datatable(x, a = 2, plot_it = FALSE)
}
\arguments{
\item{x}{The vector or single columned-dataframe containing numbers to be raised to an exponent.}

\item{a}{The exponent to raise the input values to.  \code{a = 2} by default.}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A dataframe with all inputs in one column, and their calculated powers in another.
}
\description{
Raise a set of numbers to a power and store the results as a dataframe
}
\details{
This function takes a vector of numbers as the input, and raises each number to an exponent.

This function returns a datatable with two columns. The first column, Base, contain the input numbers.
The second column, Power, contain the result of the inputs raised to the exponent.
}
\examples{
power_datatable(1:10)
power_datatable(c(2, 5, 20), 3, TRUE)
}
