% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.R
\docType{package}
\name{powers-package}
\alias{powers}
\alias{powers-package}
\title{powers: Apply powers to your vector}
\description{
Details go here
}
\author{
\strong{Maintainer}: First Last \email{b13472003@hotmail.com}

}
\keyword{internal}
