% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{powers}
\alias{powers}
\alias{powers-package}
\title{Computation of Powers and Box-Cox transformations}
\description{
this package calculate the square and cube power of a vector,
and perform Box-Cox transformations.
}
\details{
There are three functions in the package `square()`, `cube()`, and boxcox()
}
\note{
This package is a practice for STAT547 based on Vicenzo Coia's powers package. It is useful, but not intented to be a serious package
}
\author{
Santiago David
}
