% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{Linear regression between two vectors}
\usage{
regress(x, y, plot_it = FALSE)
}
\arguments{
\item{x}{The independant variable}

\item{y}{The dependant variable.}

\item{plot_it}{Display a plot of \code{x} vs \code{y} Use logical.}
}
\value{
The summary of the linear regression between \code{x} and \code{y}.
}
\description{
Linear regression between two vectors
}
\details{
This function performs a linear regression of \code{y} as a function of \code{x}, plots the results and returns a summary.
}
\examples{
a <- c(1:10)
b <- c(1:10)
regress(a,b,TRUE)
}
