% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewvec.R
\docType{data}
\name{skewvec}
\alias{skewvec}
\title{Right-skewed vector of numbers}
\format{This (and the next) is a new tag that applies for data documentation.}
\source{
Note that you shouldn't use the `@export` tag when documenting data!
}
\usage{
skewvec
}
\description{
Numbers are: -5.0 -4.0 -6.0 -2.0  0.0  1.0  2.2  1.5  1.0  2.0  1.0
4.0  2.0  2.0  1.0  0.0  0.0  0.0  1.1  3.2  4.0  3.0  2.0  9.9 10.6
16.9 21.0 23.0 27.0 39.0
}
\keyword{datasets}
