% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negtenvec.R, R/tenvec.R
\docType{data}
\name{tenvec}
\alias{tenvec}
\alias{tenvec}
\title{Integer vector from -10 to -1}
\format{This (and the next) is a new tag that applies for data documentation.}
\source{
Note that you shouldn't use the `@export` tag when documenting data!

Note that you shouldn't use the `@export` tag when documenting data!
}
\usage{
tenvec

tenvec
}
\description{
Self-explanatory!

Self-explanatory!
}
\keyword{datasets}
