% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_if_possible.R
\name{unlist_if_possible}
\alias{unlist_if_possible}
\title{Flatten a properly formatted list}
\usage{
unlist_if_possible(x)
}
\arguments{
\item{x}{List to flatten}
}
\value{
A flattened list, or an error if the list does
not contain only numeric singletons.
}
\description{
If a list contains numeric singletons, flattens the list.
}
