% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcyl.R
\name{volcyl}
\alias{volcyl}
\title{Volume of a cylinder}
\usage{
volcyl(x, y, plot_it = FALSE)
}
\arguments{
\item{x}{The radius of a cylinder.}

\item{y}{The height of a cylinder.}

\item{plot_it}{Display a plot of \code{x} vs the output. Use logical.
\code{FALSE} by default.}
}
\value{
The volume of a cylinder with radius \code{x} and height \code{y}.
}
\description{
This function calculates the volume of a cylinder.
}
\examples{
volcyl(5, 6)
volcyl(2, 7, plot_it=TRUE)
}
