% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{bc_transform}
\alias{bc_transform}
\title{Perform a Box-Cox transformation}
\usage{
bc_transform(x, lambda = 1)
}
\arguments{
\item{x}{Vector of data points}

\item{lambda}{Lambda value to be inputted manually or estimated using bc_lambda()}
}
\value{
Vector of transformed data points
}
\description{
Perform a Box-Cox transformation
}
\examples{
bc_transform(tenvec, 0)
bc_transform(-5:5, 1)

}
