% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box cox.R
\name{box_cox}
\alias{box_cox}
\title{Box cox transformation}
\usage{
box_cox(y, lambda1, lambda2, na.omit = FALSE)
}
\arguments{
\item{y}{The vector to be transformed.}

\item{lambda1}{The first parameter of the transformation, range from -5 to 5.}

\item{lambda2}{The second parameter of the transformation, lambda2 in introduced for negative
values of y, lambda2 is greater than -y.}

\item{na.omit}{Whether or not to remove NAs, the default is FALSE.}
}
\value{
A vector that is the box cox transformation of \code{y}.
}
\description{
This function perform box cox transformation.
}
\details{
This function calculate the box cox transform of a vector.


Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{box_cox} function is quite simple.
     \item The \code{box_cox} is used to transform non-normal dependent variables into a normal shape.
}
}
\examples{
box_cox(5:10,2,6)
box_cox(-5,3,2)
}
