% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_Cox.R
\name{Box_Cox}
\alias{Box_Cox}
\title{Box Cox transformation
This function perform box cox transformation}
\usage{
Box_Cox(y, lambda1, lambda2, na.omit = FALSE)
}
\arguments{
\item{y}{The vector to be transformed}

\item{lambda1}{The first parameter of the transformation in the range from -5 to 5}

\item{lambda2}{The second parameter of the transformation, lambda2 is used for
negative values of y. lambda2 must be greater than -y.}

\item{na.omit}{whether or not to remove NAs, the default is FALSE}
}
\value{
A vector that is the box cox transformation of \code{y}
}
\description{
Box Cox transformation
This function perform box cox transformation
}
\details{
This function calculates the box cox transformation of a vector
The \code{Box_Cox} is used to transform non-normal dependent variables into a normal distribution
}
\examples{
Box_Cox(1:10,2,4)
Box_Cox(-3:-1,3,2)
}
