% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.inv.R
\name{boxcox.inv}
\alias{boxcox.inv}
\title{Inverse boxcox transformation of a vector/number}
\usage{
boxcox.inv(x, lambda)
}
\arguments{
\item{x}{The vector to be inverse boxcoxed.}

\item{lambda}{The parameter of Box–Cox transformation}
}
\value{
A vector/number that is the inverse boxcox transformation of \code{x}.
}
\description{
Transforms via the Box-Cox transform.
}
\details{
We only do the boxcox transformations that only requires one input.
}
\examples{
boxcox.inv(1:10, 2)
boxcox.inv(2, 0)
}
