% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{Simplified Box-Cox Transformation}
\usage{
boxcox(x, lambda, delta, plot_it)
}
\arguments{
\item{x}{Initial data set.}

\item{lambda}{Estimate of power transformation factor to make \code{x} a normal distribution.}

\item{delta}{Amount to shift x by to ensure positive values.}

\item{plot_it}{Display a plot of \code{x} vs the output. Use logical.}
}
\value{
The vector \code{y}, which represents that transformation of \code{x} by the Box-Cox algorithm.
}
\description{
This function applies an estimate of the power transformation factor (lambda) in a deterministic power function to turn non-normal data into normal data.
However, since lamba is user-defined in this function, it may not necessarily result in a normal distribution. A more complex method, such as
maximum likelihood estimate (MLE) or Newton-Raphson, may be used to estimate the value of lamba required to achieve a Gaussian distribution, however
this is beyond the scope of this assignment.
}
