% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{Apply a box cox power transformation}
\usage{
boxcox(y = 1, lambda = 1, add_plot = FALSE)
}
\arguments{
\item{x}{The vector to be transformed.}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item boxcox (for \code{boxcox})
}
 of \code{x}.
}
\description{
This function carries out a box cox transformation
}
\details{
this does a box cox transformation
}
\examples{
boxcox(c(1:10), 2)
}
