% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{Box Cox Transformation function}
\usage{
boxcox(x, lambda, na.rm = FALSE)
}
\arguments{
\item{x}{Vector to be transformed.}

\item{lambda}{the scalar exponent for transform (1 is linear, 0 is log)}

\item{na.rm}{ask if NA's should be removed for the computation, default is FALSE}
}
\value{
A vector that is the Box-Cox transformation of \code{y}
}
\description{
This function performs a Box-Cox transformation to a vector
}
\details{
this function is useful to transform data usually to adjust them to a normal distribution
}
\examples{
boxcox(1,2,10,40)
boxcox(-10,3,5)

}
