% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{Applies a Box-Cox transformation}
\usage{
boxcox(x, lambda)
}
\arguments{
\item{x}{The vector to be transformed}

\item{lambda}{The exponent}
}
\value{
A vector that is the Box-Cox transformation of the vector \code{x}.
}
\description{
This function applies a Box-Cox transformation to any set of vectors.
}
\details{
This function performs a simple Box-Cox transformation on any integer inputted.

Box-Cox transformations are useful when:
\itemize{
     \item trying to normalizing assymetrical data.
     \item accounting for natural variance in a dataset.
}
}
\examples{
boxcox(1:10,2)
boxcox(-5,0)
}
