% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{Boxcox transformation of a vector}
\usage{
boxcox(x, lambda)
}
\arguments{
\item{x}{The vector to be boxcoxed.}

\item{lambda}{The parameter of Box–Cox transformation}
}
\value{
A vector that is the boxcox transformation of \code{x}.
}
\description{
Transforms via the Box-Cox transform.
}
\details{
We only do the boxcox transformations that only requires one input.

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{boxcox} function only requires one input.
     \item There's nothing else to say about \code{boxcox}.
}
}
\examples{
boxcox(1:10, 2)
boxcox(2, 0)
}
