% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R, R/square.R
\name{cube}
\alias{cube}
\alias{square}
\title{Cube the input}
\usage{
cube(x)

square(x)
}
\arguments{
\item{x}{Vector of numerics}

\item{x;}{Vector of numerics}
}
\value{
Cube of vector x

Square of vector x
}
\description{
Cube the input

Square the input
}
