% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R, R/square.R
\name{cube}
\alias{cube}
\alias{square}
\title{cube the input}
\usage{
cube(x)

square(x)
}
\arguments{
\item{x}{Vector of numerics}

\item{x}{Vector of numerics}
}
\value{
The vector x, cubed \code{}

The vector x, squared
}
\description{
cube the input

Square the input
}
