% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R, R/reciprocal.R, R/square.R
\name{cube}
\alias{cube}
\alias{reciprocal}
\alias{square}
\title{Apply a useful power}
\usage{
cube(x)

reciprocal(x)

square(x)
}
\arguments{
\item{x}{Vector or list of numbers.}
}
\value{
A vector (atomic, double) that is the:
\itemize{
     \item square (for \code{square})
     \item cube (for \code{cube})
     \item reciprocal/inverse (for \code{reciprocal})
}
of \code{x}.
}
\description{
Square, cube, or take the reciprocal/inverse
of a vector OR a list of numbers.
}
\examples{
cube(tenvec)
square(-5)
reciprocal(2)
square(list(1, 4, 5))
}
