% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Calculating the cube of a number}
\usage{
cube(x)
}
\arguments{
\item{x}{numeric, vector}
}
\value{
numeric, vector
}
\description{
Calculating the cube of a number
}
\examples{
cube(2)
cube(0)
cube(c(0,1,2))
}
