% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.r
\name{cube}
\alias{cube}
\title{Cubes numeric data}
\usage{
cube(x)
}
\arguments{
\item{x}{numeric, vector}
}
\value{
numeric, vector
}
\description{
Cubes numeric data (single a number, or a numeric vector)
}
\examples{
cube(5)
cube(0)
cube(c(1,2,3))

}
