% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Cube a vector}
\usage{
cube(x, na.omit = FALSE)
}
\arguments{
\item{x}{The vector to be cubed.}

\item{na.omit}{Whether or not to remove NAs, the default is FALSE.}
}
\value{
A vector that is the cube of \code{x}.
}
\description{
That's it -- this function just raises a vector to the power of 3.
}
\details{
This function isn't complicated.

It is built for learning purpose.
}
\examples{
cube(1:10)
cube(-5)
}
