% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Cube a vector}
\usage{
cube(x, na.omit = FALSE)
}
\arguments{
\item{x}{The vector to be cubed.}

\item{na.omit}{Whether or not to remove NAs, the default is FALSE.}
}
\value{
A vector that is the cube of \code{x}.
}
\description{
That's it -- this function just raises a vector to the power of 3.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{cube} function is quite simple.
     \item There's nothing else to say about \code{cube}.
}
}
\examples{
cube(1:10)
cube(-5)
}
