% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Cube a vector}
\usage{
cube(x)
}
\arguments{
\item{x}{The vector to be cubed.}
}
\value{
A vector that is the cube of \code{x}.
}
\description{
That's it -- this function just cubes a vector.
}
\details{
This function isn't complicated. It simply cubes the vector. The input may be a single number or a vector of numbers.
}
\examples{
cube(1:10)
cube(-5)
}
