% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{cube of numeric input}
\usage{
cube(x)
}
\arguments{
\item{x}{numeric, vector}
}
\value{
numeric, vector
}
\description{
cube numeric number or vector
}
\examples{
 cube(0)
 cube(2)
 cube(c(1,2,3))

}
