% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{This function cube a vector
That's it!}
\usage{
cube(x, na.rm = FALSE)
}
\arguments{
\item{x}{The vector to be cubed}

\item{na.rm}{indicates whether to remove NA's, default is false}
}
\value{
A vector that is the cube of \code{x}
}
\description{
This function cube a vector
That's it!
}
\details{
This function relies on the internal function pow
}
\examples{
cube(5)
cube(1:10)
}
