% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{cube}
\alias{cube}
\title{Cube(x)}
\usage{
cube(x)
}
\value{
This function outputs the cube of any integer "x" inputted.
}
\description{
This function can cube any integer x.
}
\examples{
cube(3), cube(9)
}
\author{
Zachary Sherker
}
