% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Cube of a vector}
\usage{
cube(x)
}
\arguments{
\item{x}{The number or vector to be cubed}
}
\value{
A vector that is the cube of \code{x}.
}
\description{
This function takes the cube if a number or each element of a vector.
}
\details{
This function is built upon the \code{pow} function, with the \code{a} set as 3.
}
\examples{
cube(1:10)
cube(-5)
}
