% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{logarithm}
\alias{logarithm}
\title{Take the logarithm of a vector}
\usage{
logarithm(x, base = 10, plot_it = TRUE)
}
\arguments{
\item{x}{The vector to be logged}

\item{base}{The base of the log}

\item{plot_it}{a logical indicating whether to generate a plot showing x on the x-axis and x log-transformed on the y-axis.}
}
\value{
A vector that is the log of \code{x}.

The base of the log to be applied to \code{x}

if TRUE then a plot is generated; if FALSE, no plot is generated.
}
\description{
This function takes a log of a vector. The base of the logarithm is specified by the user.
}
\details{
This function requires that \code{base} be greater than or equal to 0.
The \code{lologarithm} function is essentially a wrapper around the \code{log} function in base R.
}
\examples{
logarithm(1:10, base=2, plot_it=FALSE)
logarithm(-5, base=exp(1), plot_it=TRUE)
}
