% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_log.R
\name{my_log}
\alias{my_log}
\title{Log transformation}
\usage{
my_log(x, base, na.omit = FALSE)
}
\arguments{
\item{x}{The vector to be transformed. x>=0}

\item{base}{The base of the log transformation}

\item{na.omit}{Whether or not to remove NAs, the default is FALSE.}
}
\value{
A vector that is the log transformation of \code{y}.
}
\description{
This function perform log transformation.
}
\details{
This function calculate the log transform of a vector.

This function can only calculate the log transformation of x in the range:
base^(-100) <= x <= base^100
}
\examples{
my_log(7,2)
my_log(500:510, 4)
}
