% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negtenvec.R
\docType{data}
\name{negtenvec}
\alias{negtenvec}
\title{Integer vector from 10 to 1}
\format{Vector}
\source{
Ccreated manually by Sepehr Kamal
}
\usage{
negtenvec
}
\description{
Pretty stragith forward.
}
\details{
Integer vector with intervals of 1. Starts from 10 and goes to 1.
}
\keyword{datasets}
