% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig_operations.R
\name{polar_transform}
\alias{polar_transform}
\title{Transform cartesian coordinates into polar coordinates}
\usage{
polar_transform(x, y)
}
\arguments{
\item{x}{A number or vector}

\item{y}{A number or vector}
}
\value{
polar coordinates
}
\description{
Transform cartesian coordinates into polar coordinates
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{polar_transform} function is quite simple.
     \item There's nothing else to say about \code{polar_transform}.
}
}
\examples{
pola_transform(x=c(1,4,10), y=c(2,8,32))
pola_transform(x=c(-1,-4,-10), y=c(2,8,32))
}
