% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Apply a generic power}
\usage{
pow(x, p)
}
\arguments{
\item{x}{Vector to apply power to.}

\item{p}{Exponent}
}
\value{
The vector x raised to the power of p
}
\description{
Meant for internal use only.
}
