% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Power a vector}
\usage{
pow(x, a, na.omit = FALSE)
}
\arguments{
\item{x}{The vector to be powered.}

\item{a}{The power.}

\item{na.omit}{Whether or not to remove NAs, the default is FALSE.}
}
\value{
A vector that is the power a of \code{x}.
}
\description{
This is an internal function.
}
\details{
This function isn't complicated.

This function requires the input to be numeric,
when it is not, an "invalid input" message will appear.

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{pow} function is quite simple.
     \item There's nothing else to say about \code{pow}.
}
}
