% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Power a vector}
\usage{
pow(x, a)
}
\arguments{
\item{x}{The vector to be powered.}

\item{a}{The power.}
}
\value{
A vector that is the power a of \code{x}.
}
\description{
This is an internal function.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{pow} function is quite simple.
     \item There's nothing else to say about \code{pow}.
}
}
