% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Take a vector to arbitrary power}
\usage{
pow(x, a = 2, plot_it)
}
\arguments{
\item{x}{The vector to be exponentiated.}

\item{a}{The exponent.}

\item{plot_it}{A logical. If TRUE then a quick plot of the power function is created.}
}
\value{
A vector that is the \code{a}-th exponent of \code{x}.
}
\description{
INTERNAL FUCTION used by powers library. Takes vector to arbitrary power.
}
\details{
This function is for internal use only. Other functions in the \code{powers} library depend on this.
}
