% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Computes the specified power of the input}
\usage{
pow(x, a)
}
\arguments{
\item{x}{A number}

\item{a}{A number}
}
\value{
The \code{a} power of \code{x}
}
\description{
Computes the specified power of the input
}
\examples{
pow(2, 4)
pow(813.351, 5132.9441)
}
