% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{pow}
\alias{pow}
\title{Takes the exponent of a vector}
\usage{
pow(x, a = 2)
}
\arguments{
\item{x}{The vector to be squared.}

\item{a}{The exponent}
}
\value{
A vector that is the power of \code{x}.
}
\description{
That's it -- this function just gives you the exponent of a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{square} function is quite simple.
     \item There's nothing else to say about \code{square}.
}
}
\examples{
pow(2,2)
pow(5,3)
}
