% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Generic Power Function}
\usage{
pow(x, a, na.rm = FALSE)
}
\arguments{
\item{x}{Vector to raise to some power.}

\item{a}{Power to raise \code{x} by.}

\item{na.rm}{ask if NA's should be removed for the computation, default is FALSE}
}
\value{
The vector \code{x}, raised to the power of \code{a}.
}
\description{
This function is internal and is used by the other functions
}
