% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Generic Power Function}
\usage{
pow(x, a, plot_it = FALSE)
}
\arguments{
\item{x}{Vector to raise to some power.}

\item{a}{Power to raise \code{x} by.}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.}
}
\value{
The vector \code{x}, raised to the power of \code{a}.
}
\description{
This function raises a vector to any power. I'm making it internal,
by not \code{export}ing, because I don't want users to use it.
Why not? Because I said so!
}
\details{
This function can convert non-numeric inputs into numeric values.

It is also desinged to drop NA's, and will display an warning message if this is the case.
}
