% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{pow}
\alias{pow}
\title{general function to compute any power for a vector}
\usage{
pow(x, a = 2, plot_it = TRUE)
}
\arguments{
\item{x}{The vector to be operated on.}

\item{a}{The power}
}
\value{
A vector that is raised to the power of \code{x}.

The power to apply to the vector
}
\description{
general function to compute any power for a vector
}
\details{
This function isn't complicated.

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{pow} function is quite simple.
     \item There's nothing else to say about \code{pow}.
}
}
\examples{
pow(1:10)
pow(-5)
}
