% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Powers}
\usage{
pow(x, a)
}
\arguments{
\item{x}{The vector to be transformed (squared/cubed/etc.)}

\item{a}{The power to be calculated to, for the transformation}
}
\value{
A vector \code{x} that is raised to the power of \code{a}.
}
\description{
This function calculates the power of a vector
}
\details{
This is a simple function that calculates powers, which is a number multiplied by itself a certain amount of times.

For example, 2 to the power of 4 is 2 multiplied by 2 and repeated 4 times, the result of which is 16.

This function also returns an informative error when a non-numeric input is given.
}
\examples{
pow(-2, 4)
pow(1:10, 4)
}
