% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Calculate the power of a vector}
\usage{
pow(x, a)
}
\arguments{
\item{x}{The vector to be squared/cubed/to the power of whatever.}
}
\value{
A vector that is the power of \code{x}.
}
\description{
That's it -- this function just calculates to the power of "a" for a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{square} function is quite simple.
     \item There's nothing else to say about \code{square}.
}
}
\examples{
pow(1:10, 2)
pow(-5, 3)
}
