% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise.R
\name{pow}
\alias{pow}
\title{Apply a generic power}
\usage{
pow(x, to)
}
\arguments{
\item{x}{Vector or list of numbers.}

\item{to}{Power to raise \code{x} by.}
}
\value{
A vector (atomic, double) of the input
raised to the specified power.
}
\description{
This function raises a vector or a list of numbers
to any power.
}
\examples{
raise(1:4, to = 2)
raise(2, to = 1:4)
raise(as.list(1:4), to = 2)
}
