% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.R
\docType{package}
\name{powers-package}
\alias{powers}
\alias{powers-package}
\title{powers: What the Package Does (Title Case)}
\description{
This is a toy package that calculates powers of numbers. This package is being developed for the STAT547M course at the University of British Columbia.
}
\keyword{internal}
