% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.R
\docType{package}
\name{powers-package}
\alias{powers}
\alias{powers-package}
\title{powers: Apply powers to your vectors}
\description{
RACHEL CATHERINE
}
\details{
Details go here.
}
\author{
\strong{Maintainer}: Joe Developer \email{Joe.Developer@some.domain.net}

Authors:
\itemize{
  \item Pat Developer
}

Other contributors:
\itemize{
  \item A. User \email{A.User@whereever.net} [contributor]
}

}
\keyword{internal}
