% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.R
\docType{package}
\name{powers-package}
\alias{powers}
\alias{powers-package}
\title{powers: Apply powers to your vectors}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Details go here, create this file with `use_package_doc()` in console
}
\note{
to add tests use `use_testthat()` in console
}
\author{
\strong{Maintainer}: Stefanie Lumnitz \email{stefanie.lumnitz@ubc.ca}

}
\keyword{internal}
