% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{powers-package}
\alias{powers}
\alias{powers-package}
\title{powers: Convenient Computation of Powers}
\description{
Are you tired of using the power operator in R?
    Use this package to call functions that apply common powers
    to your vectors.
}
\author{
\strong{Maintainer}: Vincenzo Coia \email{vincen.coia@stat.ubc.ca}

}
\keyword{internal}
