% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.R
\docType{package}
\name{powers-package}
\alias{powers}
\alias{powers-package}
\title{powers: Apply powers to vectors}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Details go here
}
\author{
\strong{Maintainer}: Tanvi Mittal \email{tnvimital@gmail.com}

}
\keyword{internal}
