% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{powers}
\alias{powers}
\alias{powers-package}
\title{Convenient Computation of Powers}
\description{
Are you tired of using the power operator, \code{^} or \code{**} in R?
Use this package to call functions that apply common powers
and transformations to your vectors.
}
\details{
There are four functions in this package: square(), cube(), box_cox() and my_log().
}
\note{
This package is for learning purpose.
}
\author{
Xinzhe Dong
}
