% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{powers}
\alias{powers}
\alias{powers-package}
\title{Convenient Computation of Powers and repetition of strings.}
\description{
Are you tired of using the power operator, \code{^} or \code{**} in R?
Use this package to call functions that apply common powers
to your vectors.
}
\details{
The functions also have the option to generate nice plots of your vectors of numbers, taken to some power.

When character data is used as an input, the functions in this package will generally repeat in input string multiple times in the output,
depending on the chosen power. Negative powers also reverse the order of strings. Character inputs do not generate plots.
}
\note{
This package isn't actually meant to be serious. It's just for
learning purposes.
}
\author{
Hayden Scheiber
}
