% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{powers}
\alias{powers}
\alias{powers-package}
\title{Computing powers.}
\description{
This package is for simple computation of powers and exponential functions.
}
\author{
Kate Johnson
}
