% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{powers}
\alias{powers}
\alias{powers-package}
\title{Convenient Computation of Powers}
\description{
Are you tired of using the power operator, \code{^} or \code{**} in R?
Use this package to call functions that apply common powers
to your vectors.
}
\note{
This package is a practice for STAT547 based on Vicenzo Coia's powers package. It is useful, but not intented to be a serious package
}
\author{
Vincenzo Coia
}
