% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{powers}
\alias{powers}
\alias{powers-package}
\title{powers.}
\arguments{
\item{x}{The vector to be squared.}
}
\value{
A vector that is the power of \code{x}.
}
\description{
powers.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{square} function is quite simple.
     \item There's nothing else to say about \code{square}.
}
}
\examples{
pow(1:10, 2)
pow(-5, 3)
}
