% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Calculating the square of a number}
\usage{
square(x)
}
\arguments{
\item{x}{numeric, vector}
}
\value{
numeric, vector
}
\description{
Calculating the square of a number
}
\examples{
square(2)
square(0)
square(c(0,1,2))
}
