% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a vector}
\usage{
square(x, na.omit = FALSE)
}
\arguments{
\item{x}{The vector to be squared.}

\item{na.omit}{Whether or not to remove NAs, the default is FALSE.}
}
\value{
A vector that is the square of \code{x}.
}
\description{
That's it -- this function just squares a vector.
}
\details{
This function isn't complicated.

It is built for learning purpose.
}
\examples{
square(1:10)
square(-5)
}
