% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a vector}
\usage{
square(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to be squared.}

\item{plot_it}{A logical switch. If TRUE then a quick plot of the power function is created.}
}
\value{
A vector that is the square of \code{x}.
}
\description{
That's it -- this function just squares a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{square} function is quite simple.
     \item There's nothing else to say about \code{square}.
}
}
\examples{
square(1:10, plot_it = TRUE)
square(-5)
}
