% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Squares the input}
\usage{
square(x)
}
\arguments{
\item{x}{A number}
}
\value{
The square of \code{x}
}
\description{
Squares the input
}
\examples{
square(4)
square(3321.6516512)
square(-845)
}
