% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\alias{cube}
\title{Raise a vector to a power}
\usage{
square(x, plot_it = FALSE)

cube(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to be modified}
}
\value{
A vector rasied to a selected power \code{x}.
}
\description{
That's it -- this function just computes a power of a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{square}, and \code{cube} functions are quite simple.
     \item There's nothing else to say about \code{square} and \code{cube}.
}
}
\examples{
square(1:10)
square(-5)
}
