% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squre_R.R
\name{square}
\alias{square}
\title{title: square of numeric input}
\usage{
square(x)
}
\arguments{
\item{x}{numeric, vector}
}
\value{
numeric, vector
}
\description{
single number or a numeric vector
}
\examples{
square(0)
square(2)
square(c(2,4,5))

}
