% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a vector}
\usage{
square(x, na.omit = FALSE)
}
\arguments{
\item{x}{The vector be squared}

\item{na.omit}{whether or not to remove NA, the default is FALSE}
}
\value{
A vector that is the square of \code{x}
}
\description{
This function just raises a vector to the power of 2
}
\examples{
square(6:10)
square(-3)
}
