% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{This function square a vector
That's it!}
\usage{
square(x, na.rm = FALSE)
}
\arguments{
\item{x}{The vector to be squared}

\item{na.rm}{indicates whether to remove NA's, default is false}
}
\value{
A vector that is the squared of \code{x}
}
\description{
This function square a vector
That's it!
}
\details{
This function relies on the internal function pow
}
\examples{
square(5)
square(1:10)
}
