% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers.R
\name{square}
\alias{square}
\title{Square a single value or vector}
\usage{
square(x)
}
\arguments{
\item{x, }{numeric}
}
\value{
numeric, vector
}
\description{
Squares a single numeric value or a numeric vector
}
\examples{
square(5)
square(c(4, 5, 6))
}
