% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square(x)}
\usage{
square(x)
}
\value{
The output of this function is the square of any value "x" that is input.
}
\description{
This function allows you to square any integer.
}
\examples{
square(2), square(4)
}
\author{
Zachary Sherker
}
