% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a vector}
\usage{
square(x)
}
\arguments{
\item{x}{The vector to be squared.}
}
\value{
A vector that is the square of \code{x}.
}
\description{
This function squares each element of a vector.
}
\details{
This function is built upon the \code{pow} function, with the \code{a} set as 2.
}
\examples{
square(1:10)
square(-5)
}
