% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Title of the Function}
\usage{
square(x)
}
\arguments{
\item{x}{Description of the first argument.}
}
\value{
Description of the return value.
}
\description{
A short description of what the function does OK.
}
\examples{
square(3)

}
