% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a vector}
\usage{
square(x)
}
\arguments{
\item{x}{The vector to be squared.}
}
\value{
A vector that is the square of \code{x}.
}
\description{
That's it -- this function just squares a vector!
}
\details{
A number multiplied by itself is called "squared" or "to the power of 2."

This function utilizes the pow() function in the same package to calculate squares of numbers.

This function also returns an informative error when a non-numeric input is given.
}
\examples{
square(1:10)
square(-5)
}
