% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\alias{cube}
\title{Square or Cube a vector}
\usage{
square(x, plot_it = FALSE)

cube(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to be squared or cubed.}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the square (for \code{square}) or cube (for \code{cube}) of \code{x}.
}
\description{
That's it -- these functions just square or cube a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item These functions are quite simple.
     \item There's nothing else to say about these functions.
}

If you really want to see more, check out the internal \code{\link{pow}}
function that these functions depend on.
}
\examples{
cube(1:10)
square(-5)
}
