% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Squares numeric data}
\usage{
square(x)
}
\arguments{
\item{x}{numeric, vector}
}
\value{
numeric, vector
}
\description{
Squares numeric data (single a number, or a numeric vector)
}
\examples{
square(5)
square(0)
square(c(1,2,3))

}
