% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tendf}
\alias{tendf}
\title{Dataframe from 1 to 10}
\format{What format does you data take? Integer vector.}
\source{
Where did the data come from?
}
\usage{
tendf
}
\description{
Self-explanatory!
}
\keyword{datasets}
