% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenvec.R
\docType{data}
\name{tenvec}
\alias{tenvec}
\title{Integer vector from 1 to 10}
\format{Numeric vector.}
\source{
This data was built from the R function call `seq(1,10,1)`.
}
\usage{
tenvec
}
\description{
Self-explanatory. It is simply a vector of numbers from 1 to 10, useful for testing the powers library.
}
\seealso{
\code{\link{pwr}}, \code{\link{square}}, \code{\link{cube}}, \code{\link{reciprocal}}, \code{\link{negtenvec}}
}
\keyword{datasets}
