% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenvec.R
\docType{data}
\name{tenvec}
\alias{tenvec}
\title{Integer vector from 1 to 10}
\format{What format does you data take? Integer vector.}
\source{
Where did the data come from?
}
\usage{
tenvec
}
\description{
Self-explanatory!
}
\keyword{datasets}
