% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenvec.R
\docType{data}
\name{tenvec}
\alias{tenvec}
\title{Integer vector from 1 to 10}
\format{Vector}
\source{
Ccreated manually by Sepehr Kamal
}
\usage{
tenvec
}
\description{
Pretty stragith forward.
}
\details{
Integer vector with intervals of 1. Starts from 1 and goes to 10.
}
\keyword{datasets}
