% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenvec.R
\docType{data}
\name{tenvec}
\alias{tenvec}
\title{Integer vector from 1 to 10}
\format{This (and the next) is a new tag that applies for data documentation.}
\source{
Note that you shouldn't use the `@export` tag when documenting data!
}
\usage{
tenvec
}
\description{
Self-explanatory!
}
\keyword{datasets}
